requirement.txt
# Core dependencies
langgraph>=0.0.20
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0

# For Streamlit UI (to be added later)
streamlit>=1.28.0

# Database (optional - for SQLite testing)
sqlalchemy>=2.0.0

# Testing (optional)
pytest>=7.4.0